+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is a history of the changes made to idearium-lib.
 
+## Unreleased
+
+- Updated apm to only log `unhandledRejection` and rethrow the error without forcing a restart. By default the error will still get captured by `apm.handleUncaughtExceptions` and restart the process.
+
 ## v1.0.0-alpha.33 (18 May 2018)
 
 - Updated apm to use environment variables only.